+++ ccxt.php
@@ -44,7 +44,7 @@ class DDoSProtection       extends NetworkError  {}
 class RequestTimeout       extends NetworkError  {}
 class ExchangeNotAvailable extends NetworkError  {}
 
-$version = '1.6.102';
+$version = '1.6.105';
 
 $curl_errors = array (
     0 => 'CURLE_OK',