+++ ccxt/async/exchanges.py
@@ -2065,7 +2065,7 @@ class bitfinex (Exchange):
 
     async def cancel_order(self, id):
         await self.load_markets()
-        return self.privatePostOrderCancel({'order_id': id})
+        return self.privatePostOrderCancel({'order_id': int(id)})
 
     def parse_order(self, order, market=None):
         side = order['side']
@@ -2090,7 +2090,7 @@ class bitfinex (Exchange):
         exchange = orderType.find('exchange ') >= 0
         if exchange:
             prefix, orderType = order['type'].split(' ')
-        timestamp = order['timestamp'] * 1000
+        timestamp = int(float(order['timestamp']) * 1000)
         result = {
             'info': order,
             'id': order['id'],