+++ ccxt/exchanges.py
@@ -2146,7 +2146,7 @@ class bitfinex (Exchange):
 
     def cancel_order(self, id):
         self.load_markets()
-        return self.privatePostOrderCancel({'order_id': id})
+        return self.privatePostOrderCancel({'order_id': int(id)})
 
     def parse_order(self, order, market=None):
         side = order['side']
@@ -2171,7 +2171,7 @@ class bitfinex (Exchange):
         exchange = orderType.find('exchange ') >= 0
         if exchange:
             prefix, orderType = order['type'].split(' ')
-        timestamp = order['timestamp'] * 1000
+        timestamp = int(float(order['timestamp']) * 1000)
         result = {
             'info': order,
             'id': order['id'],